work on party tests
