+++ __tests__/unit/party.model.unit.test.js
@@ -1,36 +1,15 @@
-const b = require('../../build/party')
 const m = require('../../models/party')
 
-const setWorldStartDate = (options) => {
-    const objects = require('../../generic/objects')
-    const { copyObject } = require('../../lib/utils')
-    const date = copyObject(objects.date)
-    date.year = 0
-    date.month = 2
-    date.day = 1
-    return date
-} 
+jest.mock('../../persistance/commandQueue', () => {
+    executeCommands: async (command,data) => {}
+})
 
-describe('party unit tests', () => {
-    test('build should be instance of a funtion', () => {
-        expect(b.build).toBeInstanceOf(Function)
-    })
-    test('build should return a party with of size', () => {
-        const partySize = 6
-        const actual = b.build({
-            partySize,
-            date: setWorldStartDate()
-        })
-        expect(actual.members.length).toBe(partySize)
-        expect(actual.food).toBeGreaterThan(actual.members.length)
-        expect(actual.members[0].id.length).toBeGreaterThan(0)
-        expect(actual.members[0].isAlive).toBeTruthy()
-        expect(actual.members[0].health).toBeGreaterThan(0)
-        expect(actual.members[0].health).toBe(actual.members[0].maxHealth)
-    })
+
+describe('party.model.unit.test', () => {
     test('isInDwellng should be instance of a funtion', () => {
         expect(m.isInDwelling).toBeInstanceOf(Function)
     })
+    
     test('isInDwellng should return true when party is in dwelling', async () => {
         const outputMock = {
             print: () => {}
@@ -63,6 +42,7 @@ describe('party unit tests', () => {
         const actual = await m.isInDwelling(worldMock, partyMock, outputMock)
         expect(actual).toBeTruthy()
     })
+    
     test('isInDwellng should return false when party is not in dwelling', () => {
         const outputMock = {
             print: () => {}
@@ -95,6 +75,7 @@ describe('party unit tests', () => {
         const actual = m.isInDwelling(worldMock, partyMock, outputMock)
         expect(actual).toBeFalsy()
     })
+    
     test('number of alive characters in party', () => {
         const partyMock = {
             id: 'abc',
@@ -121,5 +102,5 @@ describe('party unit tests', () => {
         const actual = m.noOfAliveMembers(partyMock)
         expect(actual).toBe(2)
     })
-
 })
+
